Update class, member and function names to official names
Furthermore, fixed some oversights and mistakes.

All functions now have the same length and stack size as the original, except for a few listed below. (Some match completely now, while others use different registers from the original)

Functions I could not "match" yet (as in, they are not the same length as the original function or do not allocate the same stack size):
- Renderer::MakeConnectionStripeAttributeBlockCore(): numDrawVertex portion mismatches. It is incremented by 2 in a loop in my code. The original is able to optimize this to count * 2, which I am not able to trigger here, for some reason.
- EmitterComplexCalc::CalcChildParticle() the portion of storing copies of the registers in the stack at the beginning of the function is weirdly different, but stack size matches.
- EmitterComplexCalc::CalcComplex() matching function length but not allocated stack size.
- EmitterSimpleCalc::ApplyAnim() extra "or r5, r4, r4" after first "_f_to_ll()" in else block.
- EmitterCalc::Calc{Simple|Complex|Child}ParticleBehavior(): emitter frameRate is reloaded almost every time it's used, but I can't replicate that exactly.