Update class, member and function names to official names
Furthermore, fixed some oversights and mistakes.

All functions now 100% match the original, except for a few listed below.

Functions I could not "match" yet (as in, they are not the same length as the original function or do not allocate the same stack size):
- EmitterInstance::EmitCommon()
- Renderer::MakeStripeAttributeBlock()
- System::Initialize()
- System::AddSortBuffer()
- System::RenderSortBuffer()
- TemporaryBuffer::Alloc()
- EmitterSimpleCalc::MakeEmitterUniformBlock()
- Resource::Initialize()

However, the differences are very minor and should not affect the "correctness" of the functions. (I am certain some of these only don't match because I probably still didn't get Nintendo's build flags fully replicated.)